Séparation du code, un qui gère l'actualisation des menus
